[REF] awesome_dashboard: move control panel buttons to Layout slot
- Moved dashboard buttons into Layout's dedicated slot (t-set-slot="layout-buttons")
- Removed unsupported className prop usage; nested buttons in display instead
- Created a dedicated file for the buttons component
- Set default prop values where applicable

This refactors the dashboard layout for proper slot usage
and improves component organization.
