+++ awesome_dashboard/static/src/js/awesome_dashboard.js
@@ -1,0 +1,21 @@
+import { Component, onWillStart, useState } from "@odoo/owl";
+import { useService } from "@web/core/utils/hooks";
+import { loadStatistics } from "@awesome_dashboard/services/statistics_service";
+import { DashboardItem } from "./dashboard_item";
+
+export class AwesomeDashboard extends Component {
+    static template = "awesome_dashboard.AwesomeDashboard";
+    static components = { DashboardItem };
+
+    setup() {
+        this.action = useService("action");
+
+        this.state = useState({
+            stats: null,
+        });
+
+        onWillStart(async () => {
+            this.state.stats = await loadStatistics();
+        });
+    }
+}
